+++ src/schemas/database/project/ductType_v0.ts
@@ -7,7 +7,7 @@ export const DuctTypeDataSchema_v0 = z.object({
     maxPressureDropRate: z.number().optional(),
     maxVelocity: z.number().optional(),
     name: z.string().optional(),
-    timestamp: z.number().optional(),
+    timestamp: z.number().optional().meta({ disableUserWrite: true }),
 })
 
 export type DuctTypeData_v0 = z.infer<typeof DuctTypeDataSchema_v0>