+++ src/schemas/database/project/registerType_v0.ts
@@ -15,6 +15,6 @@ export const RegisterTypeDataSchema_v0 = z.object({
     ...RegisterSpecificDataSchema_v0.shape,
     defaultType: DefaultRegisterTypeSchema_v0.optional(),
     name: z.string().optional(),
-    timestamp: z.number().optional(),
+    timestamp: z.number().optional().meta({ disableUserWrite: true }),
 })
 export type RegisterTypeData_v0 = z.infer<typeof RegisterTypeDataSchema_v0>