+++ src/schemas/database/project/roofType_v0.ts
@@ -5,7 +5,7 @@ export const RoofTypeDataSchema_v0 = z.object({
     color: z.string().optional(),
     name: z.string().optional(),
     surfaceAbsorptance: z.number().optional(),
-    timestamp: z.number().optional(),
+    timestamp: z.number().optional().meta({ disableUserWrite: true }),
     uValue: z.number().optional(),
     unconditionedCoolingTempF: z.number().optional(),
     unconditionedHeatingTempF: z.number().optional(),