+++ src/schemas/index.ts
@@ -1,3 +1,5 @@
+import './zodMeta'
+
 export * from './api'
 export * from './database'
 export * from './getJSONSchema'